Style SVGs differently
I just realized that I may not be able to use the SVG sprite after all.
The problem is that I need to add some event handlers to the sprites, if I want
to use them as buttons. But I'm using dangerouslysetinnerhtml because
React doesn't support all SVG attributes yet, so I can't really have events.

Or can I?

One thing I do know is that I'm confused as f*&#.